+++ packages/client.test.ontology/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @osdk/client.test.ontology
 
+## 2.8.0-beta.20
+
+### Patch Changes
+
+- @osdk/api@2.8.0-beta.20
+
+## 2.8.0-beta.19
+
+### Patch Changes
+
+- @osdk/api@2.8.0-beta.19
+
 ## 2.8.0-beta.18
 
 ### Patch Changes