+++ packages/create-app.template.tutorial-todo-aip-app.beta/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @osdk/create-app.template.tutorial-todo-aip-app.beta
 
+## 2.8.0-beta.20
+
+## 2.8.0-beta.19
+
 ## 2.8.0-beta.18
 
 ## 2.8.0-beta.17